+++ src/ska_tmc_centralnode/manager/event_manager.py
@@ -49,9 +49,11 @@ class CentralNodeEventManager(EventManager):
     ) -> None:
         """
         It handles the adminMode events of csp controller device.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["adminMode"].put(event)
 
@@ -60,9 +62,11 @@ class CentralNodeEventManager(EventManager):
     ) -> None:
         """
         It handles the adminMode events of sdp controller device.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["adminMode"].put(event)
 
@@ -71,9 +75,11 @@ class CentralNodeEventManager(EventManager):
     ) -> None:
         """
         It handles the adminMode events of mccs controller device.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["adminMode"].put(event)
 
@@ -81,9 +87,11 @@ class CentralNodeEventManager(EventManager):
         """
         It handles the health state events of subsystem's
         controller and subarray devices.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["healthState"].put(event)
 
@@ -91,9 +99,11 @@ class CentralNodeEventManager(EventManager):
         """
         It handles the admin mode events of subsystem's
         controller and subaray devices.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["adminMode"].put(event)
 
@@ -101,27 +111,33 @@ class CentralNodeEventManager(EventManager):
         """
         It handles the state events of subsystem's
         controller and subaray devices.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["state"].put(event)
 
     def assignedresources_event_callback(self, event: tango.EventData) -> None:
         """
         Handles assigned Resources event.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["assignedResources"].put(event)
 
     def obsstate_event_callback(self, event: tango.EventData) -> None:
         """
         Handles observation state event.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["obsState"].put(event)
 
@@ -129,9 +145,11 @@ class CentralNodeEventManager(EventManager):
         """
         Method to handle and update the latest
         value of dishMode attribute.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["dishMode"].put(event)
 
@@ -141,9 +159,11 @@ class CentralNodeEventManager(EventManager):
         """
         Method to handle kValueValidationResult from dish
         leaf node.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["kValueValidationResult"].put(
             event
@@ -154,9 +174,11 @@ class CentralNodeEventManager(EventManager):
     ) -> None:
         """
         Handle DishVccMapValidationResult change event.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["DishVccMapValidationResult"].put(
             event
@@ -171,7 +193,8 @@ class CentralNodeEventManager(EventManager):
 
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["isSubsystemAvailable"].put(event)
 
@@ -184,7 +207,8 @@ class CentralNodeEventManager(EventManager):
 
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         self._component_manager.event_queue["isSubarrayAvailable"].put(event)
 
@@ -195,9 +219,11 @@ class CentralNodeEventManager(EventManager):
         Callback for longRunningCommandResult attribute.
 
         Delegates to specific handlers based on device type.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         if MID_CSP_MLN_DEVICE in event.attr_name:
             self._handle_load_dish_cfg_result_callback(event)
@@ -211,9 +237,11 @@ class CentralNodeEventManager(EventManager):
     ) -> None:
         """
         Special handler for LoadDishCfg and SetKValue result events.
+
         Args:
             event_data (tango.EventType.CHANGE_EVENT): to flag the
-            change in event.
+                change in event.
+
         """
         if getattr(event, "attr_value", False):
             self._component_manager.event_queue["loadDishConfigResult"].put(